+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - 2024-12-19
+
+## Added
+
+- Add `TwigNotFoundSlimMiddleware` that can be used to let Slim set correct
+  status code (`404`) when Twig template is not found
+
 ## [0.12.1] - 2024-12-11
 
 ### Fixed